requirements.txt
Scrapy==2.11.2
python-dateutil==2.9.0
scrapy_playwright==0.0.41